most settings changed to lua.
git-svn-id: svn://127.0.0.1@833 2ece6d47-a862-c445-b01b-7049066ce7d0

+++ dwindow/theater_zhuzhu.cpp
@@ -215,9 +215,9 @@ static INT_PTR CALLBACK threater_countrol_proc( HWND hDlg, UINT msg, WPARAM wPar
 			Button_SetCheck (GetDlgItem(hDlg, ID_STEREO+mode), TRUE);
 
 			// video device
-			if(player->m_output_mode == hd3d)
+			if((int)player->m_output_mode == hd3d)
 				Button_SetCheck (GetDlgItem(hDlg, IDC_HD3D), TRUE);
-			if(player->m_output_mode == dual_window)
+			if((int)player->m_output_mode == (int)dual_window)
 				Button_SetCheck (GetDlgItem(hDlg, IDC_DUALPROJECTOR), TRUE);
 
 			// monitors
